Add error case for infinite recursion with circular refs
I hit a particular edge case when working on PearlBee (a Dancer2 blog
engine). I added a component which used a circular reference (it used
the application config, and then it went to the config itself). While
maybe it's not a great idea to use circular references anyway, it made
Dancer2 go into infinite recursion whenever an exception happened inside
Dancer2.

To trigger the issue, the following has to happen:

- have a circular reference on the config or session;
- have show_errors enabled;
- have an exception somewhere in the stack;

The last item is a bit weird, because a simple `die` or throwing
Dancer2::Core::Error don't cause it. But, for example, if the template
engine dies, or if DBIx::Class dies, then the problem appears. It might
be related to the exception being blessed? I tried to find out why, but
was unable to.

So in my test, I just used a template not found.

Show_errors will make Dancer2::Core::Error dump the stack trace, the
settings of the app, the session, and the request environment. And
before dumping, it will traverse the data and try to sensor things that
look sensitive. When there are circular refs, it will traverse forever.