Fix infinite recursion with circular refs
The previous commit added a test case for this issue. This commit uses a
`$visited` hashref to the _censor subroutine, so that it keeps track of
the data structures it visited.

Whenever _censor would recurse, it used to copy the data structure like
this:

    $hash->{$key} = { %{ $hash->{$key} } };

That would change the address of the data, and make it harder for us to
keep track of visited structures. So instead, I used Clone to make a
deep copy of the entire structure on one go, and then I'm free to remove
the sensitive data as I please, with no fear of interfering with the
original.