Generate 500 response on body deserialization fail
Removing HTTP::Body caused the issue reported in #794 to reoccur.
(As HTTP::Body would process JSON content and throw an error).

As Role::Serializer wraps the deserialization in an eval and always
returns (undef on failure), to catch an error occuring; this locally
replaces the serializer log_cb so we get a signal of the failure and
die if necessary.

Its, umm, well, messy and should be considered a workaround (at best).