Cookie undef? Don't try to create object.
This is in response to a "blead breaks CPAN" ticket filed in the Perl 5 bug
queue at https://rt.perl.org/Ticket/Display.html?id=132950, wherein
"Modification of non-creatable hash value attempted" exceptions were reported.

We shouldn't try to create a Dancer2::Core::Cookie object if the cookie's
value cannot be determined.

Note that this patch causes test failures in t/request.t.  But that appears to
have to do with how the test is written.  The patch clears up the exception
and enables t/request.t to complete, albeit not with a PASS.