Handle XSCookies to return an arrayref of values
Both Cookie::Baker and Cookie::Baker::XS return a string for the cookie
value, for which we then split on `&` for handling multiple values. The
alternative XS implementation for cookie crushing, HTTP::XSCookies does
the split, returning an arrayref.

Update the cookie object construction code to accept a string (which we
split), or an arrayref of values (which we leave alone). Allows use of
Cookie::Baker(::XS) or HTTP::XSCookies with minimal code to maintain.

Closes #1435.