Config and code for not adding default middleware
If `no_default_middleware` is true, do NOT wrap the app in
  * Plack::Middleware::FixMissingBodyInRedirect
  * Plack::Middleware::Head

This allows anyone who needs to manipulate response headers based on the
response body in middleware to do so, before middleware layers like
Plack::Middleware::Head are applied.

These default middleware layers can  then be applied as part of the app's
PSGI stack (not part of Dancer2's internals).

Tests included!