don't decode already decoded date strings
This blew up before in certain locales, such as ru_RU.UTF-8 or ko_KR.UTF-8 or he_IL.UTF-8. I've talked it through at https://stackoverflow.com/q/55240390/1331451, where we concluded that the problem is that it's decoding things that are already decoded, and therefore dies. As Encode does not show the error message by default, this just silently fails.