Catch request deserialization failures
When a request that should be deserialized has bad content (such as
invalid JSON), the serializer throws an error and returns undef.
That gets caught and request object creation dies, which is then
handled in the App's psgi sub resulting in a 500 error.

When deserialisation fails, one typically does NOT want the client
repeating the request, in which case a 400 Bad Request error would be
more appropriate.

Wrap the request instantiation in an eval (with shim), so that we can
throw an error on deserialization failure.

By throwing an error in this way, we can send a serialized response
(another small win) rather than the plain text 500 error. PLUS ..
one can use the existing core.error.* hooks to modify the error as
needed, providing a solution to #1482.