Always throw a 400 response if we can't build request
There's a grey area here between throwing a 400 Bad Request, and 500
Internal Error. As re-sending a failed request at this point is always
going to fail, send this as a 400 response; so the client doesn't try
again.

This is a change in behaviour in core; we always sent this as a 500
error before. However issues #1496, #1507 and #1056 point to a 400 error
being the better option.