Improve Path::Tiny usage:
Instead of moving from Path::Tiny back to string as early as possible,
we're keeping it for a bit longer.

The defined() and -f checks can be done with Path::Tiny too. Then we
can just call openr_raw() instead of creating another Path::Tiny object.

Once we're done with that, we can finally stringify and go back to string
form.

Eventually, we would want to make all of our internals assume on Path::Tiny.