Views are stored relative to the application
Looking for them relative to the configuration, not the application,
seems weird.

Before this change, setting the DANCER_CONFDIR environment variable
changed where the application looks for views, unless you also set
DANCER_VIEWS to compensate for that.  The documentation doesn't suggest
this and it seems unintuitive to me as views aren't a type of
configuration.

The previous behaviour seems to date back to 747df69a457a028ee.