YAML serializer helpers defer loading YAML
The core tries (hard) not to load YAML if it is not required, including
having the YAML serializer only loading YAML.pm when an object is
instanciated.

The `to_yaml` and `from_yaml` keywords call helpers that do not
instanciate the YAML serlaiizer. The Mutable serializer also uses those
helpers, causing serialization to fail .. silently.

Use Sub::Defer (which is a dependency for Moo), to load yaml then
replace itself with the previous implementation of calling serialize or
deserialize as a class method.

Added test to show how this impacts the DSL helpers. Resolves #1568.