Generate a Dockerfile for new Dancer2 apps.
This was shamelessly stolen from Mojolicious. To be fair, it has been on
my todo list for months.

When dancer2 gen is provided a -c or --docker option, a very basic
Dockerfile is generated for the new application. There should be a
cookbook recipe to coincide with this, but for now, this will suffice.