Combine cpanfiles into one with options/features.
This gives us the best of both worlds, but with a bit of repeating. It
keeps the recommends definition for XS modules (and adds in the missing
ones), and gives us a feature bundle to easily install them all.

This also allows me to remove a bunch of ugliness from dancer2 gen.