Bundle App::Cmd 0.331 for backwards compatibility.
Later App::Cmd versions require perl 5.24 (for 0.332) then 5.20 (for
0.333).

Whilst admittedly 5.20 is 7 years old, there's still quite a few people
out there using older perls, and we've always tried to support older
perls where it's practical to do so.  If we decide to increase our
minimum required perl version, it should be our decision taken carefully
and announced to our users with plenty of warning, not a sudden change.

We considered pinning to App::Cmd 0.331, but that could cause Fun for
people who'd already installed a later version etc.

So, for now, just bundle a copy of App::Cmd with Dancer2, as
Dancer2::Bundled::App::Cmd, until we decide on our long-term course of
action (which may be to replace use of App::Cmd with something else).

For #1602.