Replace File::Find with Path::Tiny.
"The journey of a thousand miles begins with a single step." We have
discussed migrating to Path::Tiny (and started, see #1264 and #1544 for
details), and this is the first concrete, discrete step in doing so. It
takes a problem of small scope (app scaffolding and File::Find) to set
the stage for the rest of the migration.

There are a few TODO items here, and if there are better ways to handle
them, I am all ears. If not, I will remove my comments when I merge.

Feedback welcome/needed!