Clarify D2::Template::Simple's role in life.
This is one thing to come out of Issue #1615 on GitHub. Got me to
thinking about why we have two template systems in core when we really
want to steer devs to other template engines.

After talking with SysPete and veryrusty, it was decided we should
update the pod here to clarify what role Dancer2::Template::Simple
fills, and to consider splitting Dancer2::Template::Tiny to another
distribution.