Don't run author/release tests on < Perl 5.20.
Travis CI fails on older Perls because of Dist::Zilla's dependencies.
App::Cmd requires Perl >= 5.20, which is why we migrated from App::Cmd
to begin with. This requirement keeps us from running author and release
tests on older Perl versions.

If this is correct, we only run author/release tests on Perl 5.20 and
above going forward.