Copy Dockerfile from the right spot.
I tested this on a different machine than I developed on prior to
release, and had some cruft that masked a bug when adding a Dockerfile
to a new application. The dockerfile is located off of the dist dir, not
off of the skel dir. This minor fix changes that.

I consider this a blocker for the release. Please review ASAP.