Revised scope, other updates
Reduced the scope of the tutorial; the goal is to help accomplish
something in a short amount of time, not to be something production
ready. To that end, the relentless removal of content continues.

Updated "Where to Go from Here" to match some of the changes recently
made to tutorial code. This guide diverges from the current state of
tutorial code, however, and needs to be updated to reflect that.