Add new hook for exceptions in hooks (#1737)
This adds a new type of hook that is called in the event of an exception
in a hook. Because of its nature it's a bit more complicated than a
normal hook, in particular because it could result in recursive calling.

Because this hook can be called in quite a variety of ways during
response handling, and because the current hook exception behavior is
quite blunt, I also wanted to make it quite flexible in terms of
controlling what happens. In particular, I wanted to be able to change
the default response in the event of an exception in a hook. The
comments in the code and the documentation should all be
self-explanatory in this regard.